user management module update

+++ database/migrations/2022_02_18_092354_create_template_indicators_table.php
@@ -29,9 +29,9 @@ class CreateTemplateIndicatorsTable extends Migration
             $table->foreignIdFor(IndicatorGroup::class);
             $table->foreignIdFor(IndicatorType::class);
             $table->foreignIdFor(IndicatorUnitOfMeasure::class);
-            $table->integer('indicator_weight')->nullable();
-            $table->integer('indicator_target')->nullable();
-            $table->integer('indicator_achivement')->nullable();
+            $table->decimal('indicator_weight', 5, 2)->nullable();
+            $table->decimal('indicator_target', 10, 2)->nullable();
+            $table->decimal('indicator_achivement', 10, 2)->nullable();
             $table->boolean('is_backlog_indicator')->default(false);
             $table->mediumText('remarks')->nullable();
             $table->dateTime('deleted_at')->nullable();
